Introduce more standard project structure
Change structure to make it closer to the standard Go module structure.

Change module definition file to reflect repo location.

Retool Dockerfile to build from this new structure

To compile the minimum Go version had to be bumped from 1.18 to 1.21, as github.com/prometheus/client_golang (as downloaded by the original Dockerfile) required it to build.